testei +/-, resolvi os bugs q me apareceram ao correr as funções, no idea se funciona msm, qlqr coisa liguem me

+++ Stack.c
@@ -34,12 +34,12 @@ void add(Stack *s, uint64_t value)
     s -> top = newNode;
 }
 
-uint8_t sPop(Stack *s)
+uint64_t sPop(Stack *s)
 {
     if(s->size==0) return 0;
     s->size--;
     Node* oldNode = s->top;
-    uint8_t value = oldNode -> value;
+    uint64_t value = oldNode -> value;
     s -> top = oldNode -> prev;
     free(oldNode);
     return value;
@@ -54,11 +54,11 @@ void printStack(Stack *s)
 {
     if(s -> size == 0) {return; }
     Node* node = s -> top;
-    printf("\n%d,", node -> value);
+    printf("\n%lld,", node -> value);
     while(node -> prev != NULL)
     {
         node = node -> prev;
-        printf("%d", node -> value);
+        printf("%lld", node -> value);
         if(node -> prev != NULL) {printf(","); }
     }
 }
